Fix side effects in RequestHeadersConstraint
The constructor in RequestHeadersConstraint takes in a array of header
parameters from the Swagger definition and modifies them by calling
`unset($headerParameter->name);` and `unset($headerParameter->required);`.
This changes the definition and can cause subsequent assertions to fail.

The RequestQueryConstraint constructor calls
`$queryParameter = clone $queryParameter;` to prevent this, so I just did
the same here.
                                                                                                                                        ?>