DELETE Request with HTTP 204 Response
In case of a DELETE Request with a 204 Response the Content-Type is not set. Because of string type required for assertResponseMedia and the default of null from $response->headers->get() it fails. So from my perspective it was easiest to change the default return type to empty string to fulfill the required parameter type.